+++ backend/controllers/auth/main.go
@@ -1,0 +1,13 @@
+package auth
+
+import (
+	mailservice "backend/service/mail_service"
+	"database/sql"
+)
+
+type Authcontroller struct {
+	MyDB       *sql.DB
+	SessionKey string
+	MailInfo   mailservice.MailRequest
+	SiteUrl    string
+}
